+++ src/main/java/example/website/BusinessInterfaces/UserController.java
@@ -1,12 +1,11 @@
-package example.website.Controller;
+package example.website.BusinessInterfaces;
 
 import example.website.Exception.ResourceNotFoundException;
-import example.website.Model.Comment;
-import example.website.Model.Post;
-import example.website.Model.User;
-import example.website.Model.UserProfile;
-import example.website.Service.PostService;
-import example.website.Service.UserService;
+import example.website.Common.Post;
+import example.website.Common.User;
+import example.website.Common.UserProfile;
+import example.website.BusinessProcesses.PostService;
+import example.website.BusinessProcesses.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
