+++ src/main/java/example/website/Common/AuditModel.java
@@ -1,4 +1,4 @@
-package example.website.Model;
+package example.website.Common;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import org.springframework.data.annotation.CreatedDate;
