+++ src/main/java/example/website/DataAccess/PostRepository.java
@@ -1,7 +1,6 @@
-package example.website.Repository;
+package example.website.DataAccess;
 
-import example.website.Model.Comment;
-import example.website.Model.Post;
+import example.website.Common.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
