+++ src/main/java/example/website/DataAccess/UserProfileRepository.java
@@ -1,7 +1,7 @@
-package example.website.Repository;
+package example.website.DataAccess;
 
 
-import example.website.Model.UserProfile;
+import example.website.Common.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
