+++ src/main/java/example/website/DataAccess/UserRepository.java
@@ -1,6 +1,6 @@
-package example.website.Repository;
+package example.website.DataAccess;
 
-import example.website.Model.User;
+import example.website.Common.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
