+++ src/main/resources/application.properties
@@ -1,6 +1,9 @@
-#h2 database
-#spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
-spring.jpa.hibernate.ddl-auto= update
-spring.h2.console.enabled=true
+spring.datasource.url=jdbc:oracle:thin:@localhost:1521/XEPDB1
+spring.datasource.username=PSHARE
+spring.datasource.password=Orcl1234
+spring.datasource.driver-class-name=oracle.jdbc.OracleDriver
+# Hibernate ddl auto (create, create-drop, validate, update)
+#For deployment validate is better than update
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.OracleDialect
